moved ce functions to extra ce python file
